+++ biweeklybudget/vendored/install_vendored.sh
@@ -1,11 +1,5 @@
 #!/bin/bash
 
-# to incorporate: https://github.com/Jaymon/wishlist/pull/8
-# licensed under GPL v2
-pip install --upgrade --target . --no-deps git+https://github.com/jantman/wishlist.git@c362e59b9388f65bfa39a8c807af04f6041265b3#egg=wishlist
-curl -o wishlist/LICENSE.txt https://raw.githubusercontent.com/jantman/wishlist/c362e59b9388f65bfa39a8c807af04f6041265b3/LICENSE.txt
-curl -o wishlist/setup.py.src https://raw.githubusercontent.com/jantman/wishlist/c362e59b9388f65bfa39a8c807af04f6041265b3/setup.py
-
 # to incorporate: https://github.com/captin411/ofxclient/pull/47
 # licensed under the MIT License
 pip install --upgrade --target . --no-deps git+https://github.com/jantman/ofxclient.git@afa8d2175483bf4f50632179f434021782f49d9c#egg=ofxclient